NUSHITS (NUSHITS)
NUSHITS is a Crypto/Token that aims to encourage the adoption of digital assets in Indonesia, the country with the largest economy in Southeast Asia. Each NUSHITS will be used for the development of Staking, Farm, Swap, Pool, NFT, Dex, Crowdfunding, etc. projects to the Marketplace community and online stores